Fix transaction state machine garbage collection
When a transaction is finished it should be properly
deleted. References to the pending transactions are stored in two
lists (one for source transactions and another for destination
transactions) using boost::shared_ptrs. In order for them to be
freed we have to delete the reference from the lists.

There were some situations, very specific ones, in which the MIHF could
lock when attempting to acquire the required mutexes before removing
the reference.