[PATCH 1/3]: Deprecate mihf class and odtone-mihf library
This series of patches are for removing the mihf class, and were long
overdue.

The purpose is to remove cyclic dependencies between the MIHFs objects
(service acesss controller, net sap, transaction manager,
etc...). These cyclic dependencies were solved using the singleton
pattern, but this solution is not scalable because it adds too much
complexity when inserting new objects, i.e. another net sap.

In the process also remove the odtone-mihf library. This library was
to help build other mihfs, but because of the tight coupling between
modules, it doesn't make sense to build a library and then link the
main source file. Instead just compile and link the modules into the
binary. Has an added bonus, it should help the port to the Android
platform, because it's on less library to link to.

This first batch removes the mihf.* files and updates the build system.