Add address_book class definition
This class is used for storing information on how to send messages for
several MIH entities, such as users, link saps, and peer mihfs.

The address book is implemented using an std::map in which the key is
an id (peer mihf) and the value is an address_entry struct. An
address_entry is composed of an ip, port and transport_list_enum. The
transport_list_enum is used to identify the protocol to be used for
sending messages (UDP or TCP for now).