[PATCH 1/6]: Deprecate transaction_manager class
These series of patches are for spliting the functionality of the
transaction_manager class into a class for storing the pending
transactions, and two others. One for source transactions
(transactions started when the MIHF sends a message to a peer MIHF)
and another for destination transactions (transactions started when
the MIHF receives a message from a peer MIHF).