[PATCH 2/6]: Introduce transaction_pool class
This class is where all the pending transactions are to be stored.

It also incorporates a timer mecanism that decrements the
transaction_stop_when and retransmission_when counters. This feature
(the timer) could be in a diferent object, but because we have to
acquire a mutex in order to decrement the counters, the design of such
an object would be more complex, so I left it has part of the
transaction_pool.