Add an address_book to transmit
- Removed send() method
- Removed add() method

The purpose of the transmit class was to wrap all methods for sending
messages, and was also responsible for storing information about the
subscribed MIH Users, Link SAPs and peer MIHFs.

In order to further remove tight coupling between classes, the
transmit class will only be responsible for sending local messages
(messages for local MIH Users and Link SAPs).

If the destination identifier of the message isn't in the address book
it sends it through to the message_out functor.