Split service access controller into functions
With a service access controller (SAC) as a class there is another
cyclic dependency between message_out and the SAC, because the
transaction state machines (source and destination) expect we have a
global function called process. This function returns a response to
the incoming message.

To avoid further chicken and egg problems with the SAC just split it
into functions that operate on a static map.

The functions are __not__ thread safe because I assume that callback
insertion is only made on MIHF instantiation. So there's no need to
add mutexes to the mix.

A callback is a pointer to a function that returns a boolean and
receives two pointers to mih::messages. One is the incoming message
the other pointer is the (possible) response.