Teach transaction state machines about handler_t
The state machines need to call a process message function at certain
stages. There are two options, call the sac_process_message directly
or have a reference to a function with the same signature.

I chose the second option to decouple the objects, and to facilitate
possible unit test cases with mock objects.