Make sac_dispatch a class
The sac_dispatch is called by the local udp_listener, where the MIHF
listens for local messages.

We don't want to use transaction state machine on communications
between the MIHF and the local SAPs, mostly because of the confirm
messages and timers associated with the transactions.

The transaction state machines need a process_message function that
returns true if the message processed generated a response.

To send messages from the MIHF to a local SAP we use the transmit
functor.

The sac_dispatch is implemented as a functor class that receives on
the constructor a reference to the transmit object, because the
udp_listener that calls the sac_dispatch would need to have the
reference to pass to the (previous) sac_dispatch.

The next step is to have the udp and tcp_listener receive a function
that they chould call when a message is received.