Teach SAPs about generic MIHF ID
The standard defines a multicast MIHF ID as an MIHF ID of length
zero.

To add support for multicast messages we need to change the protocol
between local SAPs and the MIHF, because we use an MIHF destination
identifier of lenght zero to identify the local MIHF as the
destination of the message.

The proposed solution is to use the MIHF ID "local-mihf".