Add broadcast Capability_Discover support to MIHF
To enable this option pass --enable_broadcast in the command line.

When a MIHF receives a broadcasted Capability_Discover.request it
should respond with a Capability_Discover.response to the MIHF that
broadcasted.

To fire a Capability_Discover.request a MIH User must send a
Capability_Discover.request with a zero length MIHF destination
Id. The MIHF then broadcasts the request to the network, and then
forwards the response to the MIH User.