Add a MIIS server example with support for RDF and Sparql queries
This example uses the Redland RDF libraries for parsing and querying
a sqlite database, other databases can be used. Check the Redland
documentation for more details.

Pass the path to the database with the --database option. An example
database will be provided.

The usual MIH User options are available (ex. listen port).

Behavior:
  - When a MIH_Get_Information.indication is received the message is
    deserialized and the Sparql queries are fetch. The queries are
    run and the results serialized and sent in a
    MIH_Get_Information.response to the MIHF.