Rewrite the TLV DSL
The tlv_fwd implementation is ugly and duplicates to much code.

The serialization was moved to the existing TLV concept classes
tlv_, tlv4_ and oui_, to remove the code duplication.

The tlv_fwd class is know split between the tlv_type_ and tlv_type_cast_
classes. One major improvement is the specialization for the optional
case, instead of current ugly hack.