[mihf/command_service] Fix message forwarding
MIH Command messages are handled by a MIH User, when the MIHF receives
one of these messages it should forward them, currently it defaults to
local MIH User 'user'.

This patch fixes a possible issue with local transaction handling,
because it's the responsability of the MIH User to send the response
with the MIH Destination identifier set to the MIH identifier that
requested the message. The MIHF then forwards the message.